+++ conftest.py
@@ -14,7 +14,6 @@ def pytest_configure(config):
     PYTEST_HEADER_MODULES['pyyaml'] = 'yaml'
     PYTEST_HEADER_MODULES['scikit-image'] = 'skimage'
     PYTEST_HEADER_MODULES['specutils'] = 'specutils'
-    PYTEST_HEADER_MODULES['spectral-cube'] = 'spectral_cube'
     PYTEST_HEADER_MODULES['asteval'] = 'asteval'
     PYTEST_HEADER_MODULES['click'] = 'click'
     PYTEST_HEADER_MODULES['echo'] = 'echo'